+++ tests/code/reverse/copy_reduced_dealloc/reverse.h
@@ -1,3 +1,3 @@
 #include "Util.h"
-long long* reverse(_DECL_1DARRAY_PARAM(ls), _DECL_DEALLOC_PARAM(ls), _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
+int64_t* reverse(_DECL_1DARRAY_PARAM(ls), _DECL_DEALLOC_PARAM(ls), _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
 int main(int argc, char** args);
